src/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-cloud-firestore>=2.14.0
google-cloud-pubsub>=2.18.0
google-cloud-secret-manager>=2.18.0
httpx>=0.25.0
PyJWT>=2.8.0
cryptography>=41.0.0
python-multipart>=0.0.6
